refactor: rename Title-styling file
