+++ src/components/containers/domains/headers/Title.styles.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from "react-native-web";
-import theme from "../../../../../theme";
+import theme from "../../../../theme";
 
 const styles = StyleSheet.create({
     text: {
